Refactor settings migration
